containerizing the logic
this commit puts the logic into a container with access to gpu
resources in order to run the training (and subsequently the
evaluation). i have not 100% validated the evaluation within the
container yet, but i feel it should work with the changes i have
made. generally, however, the training works within the docker
container.

i technically use podman as the container engine for this, with
the nvidia-runtime-container enabled so that it can access the
gpu resources. i also need to run in privileged mode to get the
container to have access to the gpu.

there were some minor modifications to the code to get replicate
the exact study (mainly, changed the jaccard similary score import
to just jaccard_score). i have gotten this working on my system in
totality, and have frozen the pip environment i had in order to
generate the requirements.txt. the dockerfile also pulls a cuda image
to gain access to gpu resources, and installs the proper version
of python to unpickle the needed files. finally, i have added scripts
in order to build and run the container and to train and evaluate
the model once inside the container.

there is still more work to be done to include the planned ablations
and additions. i also want to figure out ways to output the model
file so that training and evaluation can happen in separate containers
(ie, can i use a volume to store the model file on the host os and
load it from the host os into the evaluation container? what other
options do i have here? etc). i also am currently using the prepared
data from the github repository. i want to start working to perform
the data preparation step in a container as well, which would also
give me the benefit of potentially updating the python version since
that was primarily chosen to support loading in the pickle files
that were included in the repository. if i prepare the data separately,
i may be able to do so with newer python versions. finally, it may
be worth creating some sort of flask or api application that can
accept requests to re-enable some of the command line arguments that
made the original code more flexible.

these are all things to be done in future commits, but how far i will
get with each of these will be determined later. the most important
is getting containerization to work and starting the ablations.