data preparation logic
Previously, I was able to get the containerized logic working to
train and evaluate the model. However, this was using the pickled
data files that were already provided by the original authors to
make the replication of the study simpler. There was also code that
the authors wrote in order to prepare the data and output the pickle
files that would later be used. I extracted out the code from the
python notebook included in the repository into separate python
files to prepare the data with, and added the execution of these
python files as steps in the Docker container as part of the
overall pipeline. Now the code works fully end-to-end -- data
preparation, model training, and model evaluation. The only piece
that is not automated at the moment is downloading the data itself.
I am considering adding this step as well, but it may be tricky
since the DDI data was provided as a Dropbox download link. Anyway,
now the code is working and I can focus on ablations as the next
steps. I may try to simplify the code that I modified in order
to make it cleaner, since my modifications were quite messy as a
result of various trials and errors. However, I want to commit the
working code now as a base for any future enhancements.

There are some idiosyncrasies as a result of some of the choices I
made. For example, in the Docker file I copy over all of the files,
and then copy the data processing logic over again. This is because
I need this context to unpickle the files. While the original pickle
files were likely able to reference the original EDA code provided
by the authors, when I extracted the logic into separate files this
caused some challenges. To avoid duplicating the files themselves,
I simply copy the files into the Docker container multiple times.

I have added the full flow as part of a "pipeline" executed in a
bash script by the Dockerfile. One thing to note is that now the
grades and feedback for the project proposal have been released.
With the feedback, I learned that the full GAMENet pipeline is
already integrated into the pyhealth library. Though this effort
so far has already taken some time, I may want to spend some time
investigating the pyhealth implementation as it may even be simpler
than what I have currently -- plus it would be valuable to learn
more about! I may investigate in a future commit or a different
repository, but we will see what makes the most sense.