baseline logic
now the logic to train and evaluate the baselines has been included
in the container. this has been done rather messily and a bit
hastily, so i plan to refactor this a bit in a future commit. however,
the code in its current state represents a runnable docker container
that succeeds in the following:
- prepare the data for training purposes, including ddi graph and
  ehr data
- trains the GAMENet model on the prepared data and evaluates it
- trains the DMNC model on the prepared data and evaluatates it
- trains the LR model on the prepared data and evaluates it
- trains the LEAP model on the prepared data and evaluates it
- trains the RETAIN model on the prepared data and evaluates it

at least, my runs seem to be successful. training the dmnc model
took quite a while, ~20 hours or so on my Nvida 1080 GPU. one thing
i noticed was that only 25% of the GPU memory seemed to be used at
any one time. i wonder if this is normal or if i can somehow adapt
the data loader patterns or something to enable more data throughput
somehow ...

one of the most important things i had to do for this was to upgrade
my version of `torch`. i upgraded quite a bit, to version 1.4; after
this, the DMNC model was able to train. before that upgrade, the
model training had issues calculating the gradient for some operations
it regarded as self-assignment. anyway, now it trains though for
a long time. all of the other baselines seem to be working from
what i can tell as well.

i want to make the bash script cleaner, and add command line arguments
to evaluate whether baseline evaluation should be run. i also want
to make some of the other code cleaner as well. i can probably
abstract some of my rougher logic out into the utils file, and find
a better way to handle my paths. as i mentioned, this was a bit of
a "rough draft" attempt but since i have it working before project
draft submission i will commit it for now. hopefully in the next
commit i can make a lot of this stuff more clean and user friendly
before merging the pr.