refactor
This commit serves primarily to refactor the existing code, to
make it a bit easier to follow and user-friendly. There are a few
central changes that contribute to this.

First, I've renamed the `train_and_eval_model.sh` script to be just
`gamenet.sh`. I felt this was more accurate, since I am interested
in running the full GAMENet "pipeline" instead of just training
and eval; this includes preparing the data and training and evaluating
the baselines, along with the GAMENet model itself. Also in this
file, I have modularized the various logic into functions that can
be controlled via simple boolean flags. My intention is to add
command-line arguments and flags to this bash script so that these
flags can be set more dynamically, but for now this helps to easily
toggle which parts of the pipeline I want to enable/disable.

I also made some changes to the Dockerfile and the structure of the
container. Instead of copying the repository as-is into the container
and then duplicating certain files as necessary and dealing with
convoluted path structure, I now copy in the files I want one-by-one
into the main directory of the container. This simplifies the imports
and the copies, so that every file only needs to be copied once.
It also simplifies the path structure for the pickle and data files.
Thus, I also needed to make changes to some of the training and
evaluation code, as well as the data preparation logic.

Another change here is that I created some helpful utilities in
the utils file. This includes evaluating whether or not the TEST
variable for a given model should be true, as well as getting the
path that the pickle files should be saved to. This will allow me
to simplify any changes that are made here, so that from now on
I only need to update the path in the utilities function instead of
across the entire project.

I do still want to do a second refactor commit, where I clean up the
structure of the repository itself. I think I can move all of the
files out of the `data/prepare_data/` directory and just put them
in the `data/` directory. I can also simplify how the Dockerfile
copies the directories and files into the container to reduce the
number of steps in the build process. I also plan to add a simple
alias for docker=podman that can be commented out for anyone who
wants to use docker instead of podman. Finally, once I add a bit
of documentation to the README and different areas, I plan to merge
this PR.