further refactoring
I think my repository is finally starting to look clean enough to
merge. It was a bit of a messy path as I continued to add new things
and experiment, but I am comfortable with the current state of the
code and the organization. I think the final steps before merging
the PR and submitting the progress report will be to add some
documentation and maybe just a little bit more cleanup (for example
cleaning up and adding additional comments).

Mostly, this commit involves moving files around. I had created
a `data/prepare_data` directory that I was using because I was
nervous to touch any mission-critical files. Now that I have done
various experiments, I feel more comfortable with how everything
works. So, I was able to move the files I wanted to and organize
it a bit better than I had in my initial commits for this PR.
Pickle files now reside in their own directory, and while these files
do get copied into the container, they will get deleted if the bash
entrypoint's prepare data flag is set to be true. I also updated the
util pickle path to follow with this new organization. I also moved
the data preparation logic up to the `data/` directory itself.

I reordered a bit of the logic in the `gamenet.sh` entrypoint, just
to make things a bit easier to follow. I haven't added command-line
arguments to the bash script or anything yet, but that may come in
a later commit. For now I am comfortable with how this works, and
can always extend it later. I have verified that all of the models
seem to train successfully (or at least start to do so). After the
progress report is submitted, I will make sure to run the entire
pipeline again to verify that the full training and evaluation is
working. I have validated this for some of the smaller, easier to
train models, but not for everything.