documentation
I cleaned up some final things that I wanted to address before
finalizing this PR; this mostly comes down to removing comments in
the various python files.

Additionally, I added an alias for docker so that it refers to podman
in the build script. Now, this line can easily be commented out in
order to use docker instead, while still letting me use podman in
general. As mentioned, podman should mostly be a drop-in replacement
for docker. I have also added a cleanup script that helps me track
the commands I need to free up the resources on my system.

Finally, I have actually made updates to the README file. I have
kept the original README file in tact, so that it can be referenced
as needed. I also did not want to lose any details that the original
authors included for now. Instead, I have added my sections at the
top of the README file. This documentation shows how one can take
this code and run it as docker or podman containers leveraging the
NVIDIA Container Runtime. This allows the containers to have access
to GPU resources as needed. Now I feel I can merge the PR and tag
a release regarding the project's progress report.