containerization, retain baseline, and documentation
This commit should be the final wrap-up for this PR. I have included
a number of changes in this commit. First, I can now run the code
in podman (or docker) containers. The container seems fairly robust.
I am able to mount the data directory containing the mimic3 and mimic4
datasets thus alleviating the need for the lengthy copy of large files
into the container. This both greatly reduces build time and also
greatly reduces container size, which are ideal. The container should
now only have the bare essentials to run for the most part, though
I am wondering what alternative options I have for the base image
and for the python/pip installation. However, further improvements
here will likely have to be (indefinitely) stalled while I switch
gears to the ablations and additional components of the experiment.
Another nice feature of this containerization effort is that I can
pass arguments in at the run step to the container, which will then
be passed to the python script entrypoint. This means that I can supply
the command line arguments (like number of epochs, values for
hyperparameters, and mimic dataset to use) at runtime of the container
and have them reflected in the script's run. One thing to note is that,
for whatever reason, the logic does run faster on my host hardware
than in a container. Not sure why, but doubt I will solve this in
the scope of this final project.

I have also added the RETAIN baseline to my execution logic. This
was fairly simple, and is trained much like GAMENet. The evaluation
logic is identical, so the function there was reused. For RETAIN,
I had to pass a few additional arguments that I was able to ascertain
from the original paper. This gives me at least a base for comparison
so that I can validate some of my hypotheses. I may spend some time
looking at if I can implement other baselines, but a few of them
seemed like they were not in the pyhealth library yet. This could
indicate a potential area for me to contribute to the library, or
it could be where I have to draw the line for now until I can get
some further work done.

Finally, documentation has been added. This includes updating the
README in the root of the repository as well as adding a README in
the pyhealth directory. This README includes directions for how to
run, the flags that can be provided to the script, etc. I also have
left comments in the code that will hopefully make it fairly easy
to parse. A cleanup script has once again been provided, with some
podman commands I found helpful when cleaning up my environment.
The last piece I have added was a few results from running both
the GAMENet and RETAIN models on the MIMIC3 and the MIMIC4 datasets.
Overall, I am impressed with RETAINS performance. It comes closer
to GAMENet than I realized, even surpassing it at times. One metric
I will have to consider implementing is the DDI Rate metric, which
is one area I feel GAMENet is probably better than RETAIN for.