ablations
This is the first commit for my work with ablations. The first
ablation I have tried to implement is omitting the history component
of the data preparation task. Strangely, this seems to perform
better than the original GAMENet model... I wonder if there is
something I am missing, if there is a change I need to make to
the GAMENet model in order to get this working properly. However,
for now this is at least a start with the ablations. I also will try
to add other ablations that make sense.

This commit also contains some major code restructuring. I have made
a class that is a wrapper around the data utilities that I have, that
can group tasks and datasets together to perform the necessary
preparation and provide the data needed for training. I don't think
this is necessarily the right approach, especially since I have
created my own `Task` class when pyhealth has its own concepts of
tasks. For now, this works and helps me dynamically switch certain
parameters such as what dataset I am using and which tasks I am
running -- so I may stick with it depending on the time I have. If
I have additional time, I will try to figure out a better way to
structure the code.

I have also added some other files, such as the start to an EDA
python notebook, but that is in its very nascent stages. I have added
a `drug_rec_task.py` that has the tasks I plan to perform for the
ablations, but most of them are experimental at the moment so I
plan to clean that up. I also included results from the runs, which
contain the strange situation of GAMENet results improving when not
considering the history of the patient.