first refactor; task command line arguments
This commit contains the first refactor effort for this PR. I think
I may add a second refactor where I abstract some of the model logic
into another wrapper class. For now, I have cleaned up a lot of the
messiness of the first commit in the PR by removing old, commented
code and moving functions around.

First, I extracted the data logic into its own file so that it
would be a bit easier to manage. The gamenet.py file was getting
large and unwieldy so I figured this would help make things cleaner.
I also added the ability for the data wrapper to run custom tasks
that get passed to it, and generally made the flow here a little
bit nicer. I still don't know if this was the right abstraction to
have here, but for now it's codified the logic well enough that I
am satisfied (at least for now).

I also added some command line arguments to manage the running of
tasks. By default, only the original drug recommendation preparation
task is run. However, a user can specify any number of optional tasks
with the `-t` argument now, though at the moment the only additional
allowed task is the no history task -- the ablation introduced
earlier in this PR that computes the patient record without considering
their history of medications, procedures, and conditions. I feel
these changes should help make the development of further tasks a
bit easier. There is also the `-a` flag, which will run all of the
possible tasks for that dataset.