second refactor; model abstractions
Abstracted the model logic out into a wrapper that helps to
encapsulate the functions that are important for training and model
evaluation. This is a small refactor compared to the previous one,
but my plan for the next commits is to add a metric around the
number of drugs being recommended by each method, and to add more
ablations as needed. I have a feeling that while the GAMENet model
without history performs better than the GAMENet model with history
at the moment, that tracking this additional metric will help prove
that the history is important by showing a degraded performance
in the drug count metric. I've tried to also replicate the DDI Rate
metric in a local branch, but so far my efforts have been
unsuccessful. I may try again after thinking through it a bit more,
but since it was relatively tricky to figure out for now I have
paused that effort.