average drugs per visit
One suggestion from the project draft (progress report) was to include
details about how many drugs each model recommends on average per
visit. This is because one of the strengths of GAMENet is that it
tends to recommend a smaller set of drugs per visit, which helps
reduce the DDI Rate of the results as well. I wanted to include this
metric as another way to compare the models.

This commit has the logic for calculating that metric, and does so
for each experiment. The surprising thing to me is that the RETAIN
model seems to recommend less drugs than GAMENet for both included
tasks. I may want to revisit how exactly I am calculating this, since
it doesn't exactly match up with the experimental results from the
original paper (though admittedly it is close, so it may be correct).
I'm hoping this approach will also help me figure out the right way
to calculate the DDI Rate of each model as well.

I have also included the results of my most recent run in this commit
to keep a record of them. For example, it is worth noting that without
the patient history, the GAMENet model tends to recommend more drugs.