ddi rate metric
I realized that my attempts at recreating the DDI Rate metric were
actually (somewhat) valid; my mistake was that I misread the DDI Rate
values in the original paper, and thought my results were a magnitude
off. Instead, they are fairly close to what is presented in the paper,
so I have added this calculated back so that I can print out the
DDI rate for each method (as well as the delta compared to the
base DDI rate of 0.0777). Strangely, it seems RETAIN recommends
fewer drugs than GAMENet on average in my implementation, which does
not correspond to what is seen in the paper (or in my original
reproduction done in the first PR of this repo). However, GAMENet
seems to consistently perform better in terms of metrics like
accuracy and DDI Rate.

Additionally, I have moved some functions around from the prior
commit as a bit of cleanup. Now the DDI Rate and the Average Drugs
per Visit metrics are calculated within the model wrapper itself.

Finally, I have added some more results from these DDI Rate runs.
As mentioned, the metrics do seem a bit strange at times so I may
continue investigating as time allows. However, the due date is
fast approaching so I may have to wrap this up soon and cut a release
for the final project submission. We will have to see how far I get
in the next day or so. I plan at least one more cleanup and
documentation related commit for this PR. Depending on time there
may be a few more commits -- such as one additional ablation if I
can get it working.