procedure ablation
One of my planned ablations was to omit the procedure information
for patients to see how the various models would perform. This commit
includes both the task that prepares the drug task data without any
procedure codes, as well as an "alternative" GAMENet model that is
trained (and evaluated) on the data prepared without procedure codes.
This new model variant is called GAMENetNoProc for now, but I will
try to come up with a better name for it. This may also be where I
try to make a PR against pyhealth to provide some options to the
GAMENet model as far as whether to include procedure information
or not...but we will see.

The results here are fairly interesting. The number of drugs per
visit that were recommended in GAMENet was actually smaller on
average, but the DDI Rate was higher compared to the original
GAMENet implementation. I may try to run this again to see if that
is consistent. For now, I have included the results of my most recent
run, as well as the alt_gamenets.py file with the GAMENet variant
I described. The new drug recommendation task has also been included
in the drug_rec_task.py file, and the task has been added to the
list of allowed tasks for command line arguments.

Additionally, I made some minor updates to the README, including
documenting the flags that have been added to the script since the
last README update.