gamenet with no hist
I realized that the ablation I tried to add for no history still
maintained a dynamic memory component within the GAMENet model, so
it likely was not working the way I expected it to. I have now added
another GAMENet variant, GAMENetNoHist, which removes the Dynamic
Memory component from the GAMENetLayer. The new layer here is called
GAMENetLayerNoDM. Previous drugs are not computed and are not sent
to the forward operation of this layer, so hypothetically now both
the data preparation and the GAMENet model should not account for
the history of the patient.

I have also made a few other edits. I tried to add a flattening task
to the drug_rec_tasks to make another ablation I had planned, but
I could not get this to work as expected. I think I will have to
skip this ablation for the sake of the project at least, as it is
a bit tricky to figure out for now.

Additionally, I have moved most of my constant values into a
constants.py file, which makes it easier when I want to make changes
that affect the entire logic -- such as adding a new task. Now, I
can just change values in the constants file (such as which task
corresponds to which model) in order to add or update certain
characteristics. I have also added timing metrics to each training,
so I can see how long it takes to train each model. Some results from
some of my runs have been added here as well.