eda notebook
I've revised the EDA notebook. Now it is working again. This notebook
demonstrates the process of loading in the MIMIC data, performing
tasks on that data, and then training and evaluating the models on
that data. The final step is that it outputs the table of model
performance characteristics, which will be used in the final report
(with some minor revisions). The notebook still needs to be cleaned
up a bit, and some documentation and a 200 word summary still need
to be added. However, I am comitting it in this state as it is generally
where I want it to be. I may try to figure out if there are other
visualizations I can add in this notebook as well.

Additionally, I cleaned up some of the code from the previous commit.
For example, I have mode the training timing logic into the model
wrapper class itself, so that now it is automatically done on
training and can easily be stored and access later on when needed.
I also removed some comments and extraneous code.