dev eda notebook and fix podman
I have finalized the EDA notebook to account for the requirements
for the Notebook bonus for the final project. Now, the notebook should
fulfill those requirements, such as containing a summary of the results
and of the entire reproduction study. For some reason, I could no
longer run the notebook with the full dataset. For now, I have just
run in dev mode to get sufficient outputs. There is still the previous
commit which contains the exact outputs with actual data which is
being used in my final report.

I also update the Dockerfile and the docker run script to enable the
changes I have made in this PR to also run in the containerization
framework I introduced previously. Now, the default behavior is for
the container to run all tasks when invoked. I will probably not
make this more fancy. The python script accepts command line arguments,
and the build_and_run script can be used to see the proper way to
pass command line arguments in to the python script. If it is easy,
I may try to add some command line arguments to the bash script.
But it seems unnecessary as I can just copy the commands and run
them manually with the proper flags if needed.

Finally, I removed some references to unused code lines, especially
with regards to the flat ablation which I was unable to get to work
successfully.