fix strict models and final documentation
Previously, I only allowed very strict model classes to be passed
in as arguments to the ModelWrapper. I have made that more flexible,
and now I allow any object that is a subclass of GAMENet or RETAIN
to be passed in as the model type for the wrapper. This allows the
EDA notebook to work a bit better as well. The notebook has been
updated to reflect those changes.

Additionally, I added some final comments in the code to document
some of the aspects to make it easier to follow. I also updated the
README with a link to my video for the assignment.