final run and model checkpoints
It is recommended to store the model weights in the repository as
well to make the results shown easier to produce. I have down that
now in the pyhealth/output/ directory, which contains the best model
for each task for both RETAIN and GAMENet. I don't have anything
in the script to easily enable reading from checkpoint files, but
this should still be easily achievable by setting the load parameter
to True in several of the ModelWrapper functions. That may be a future
enhancement to make it simpler, but the project deadline is here
so I will not get to it before the due date.

Additionally, I provided a table with the results of my final run
in the README, as well as a log from that run. I also made a few
other updates to the README to make everything explicit and clear,
and to give a sense for the purpose of the GAMENet project at a
glance -- which can be understood by delving into the source paper
deeper.

Unless I see a glaring error in the README after this commit, this
will be the final commit for the assignment at least.